Made MSXDevice::getWriteCacheLine() NON-const
The original reasoning to make it const was:
  Querying a cache-line (read or write) is an optimization, it should
  never change the state of the MSXDevice, so therefor this method
  should be const.

However getWriteCacheLine() returns a NON-const pointer. And then
writing via that pointer CAN modify the internal state of the
MSXDevice. So in a way getWriteCacheLine() could be used to circumvent
const correctness (it would allow to write in a const MSXDevice).

In a few places, in the implementation of getWriteCacheLine() we
actually had to use a 'const_cast'. Getting rid of those casts is the
main motivation for this patch.

The reason why 'const_cast' was only needed in a few places rather than
most places has to do with 'deep' vs 'shallow' const. C++ const is only
a shallow check. For example if a device has a pointer to a Ram object,
then, from a const-method, returning a non-const pointer into that Ram
is fine because, from a shallow point of view, non of the members of the
device object itself are changed (the pointer to the Ram object isn't
changed). On the other hand if the Ram were implemented via a (small)
array in the device object itself, then return a non-const pointer to it
from a const-method would not be allowed. Ideally such implementation
details (holding the Ram via a pointer or directly in array) shouldn't
matter. But for technical reasons that's not how the const-checks work
in C++.