Use 'enum class' in CliComm
This continue the work from edf127cf7b (and others) to replace 'enum'
with 'enum class'.

The reason why these weren't converted yet is that these are used as
indices in an array (that no longer works with 'enum class' because
those implicitly convert to 'int'). Workaround is to use
'to_underlying()', but that makes the code very verbose.

This patch includes a new utility 'array_with_enum_index' to make this
easier.