MegaFlashRomSCCPlus: Implement address line swizzle.
I noticed the MegaFlashRomSCCPlus used Addressing::BITS_11 while it has an
8/16-bit FlashROM chip. BITS_11 should only be used for 8-bit FlashROM chips
or 16-bit ones when accessed in 16-bit mode. When accessing a dual-mode
8/16-bit FlashROM chip in 8-bit mode, one should use BITS_12.

I could only think of two possible explanations for this; either the chip was
operating in 16-bit mode with D8-D15 left unconnected, with only half of the
capacity available, or address line A0 (A-1 on datasheet) was connected to CPU
line A11 or A12 with all lines in-between shifted down. Manuel Pazos confirmed
it was the latter, specifically Flash line A0 is connected to CPU line A12.

As a consequence of this, read caching is not possible.