Experiment: use 'enum class' for MSXDirEntry::Attrib
'enum class' has some advantages over plain 'enum'. Most important:
* Better scoping.
* No implicit conversion to the underlying type (e.g. 'int').

In case of 'MSXDirEntry::Attrib' we were actually using that implicit
conversion to 'int': we want the bitwise AND and OR operations. But
other operations like arithmentic (plus, min, ...) don't make sense.

This patch is an experiment: it uses a combination of 'enum class
Attrib' (for the scoping and 'using enum' stuff) and a new struct
'AttribValue'. The latter can (implicitly) be constructed from an
'Attrib' and then only offers the operators that make sense.

My preliminary evaluation of this experiment:
* Using this new stuff looks clean.
* But it did require some (~10 lines) of boiler-plate code. Maybe that's
  not too bad?