Change return-type for EventListener::signalEvent() from int to bool
In the past the return value represented a bitmask of priorities: an
EventListener could indicate exactly which (lower priority) listeners
would still receive the current event.

Later the meaning of the return value was changed to one specific
priority: an EventListener could block the current event for listeners
starting from a specific priority level (and all lower priorities).

Though in practice all EventListeners either returned 0 (meaning don't
block the event) or returned their own priority + 1 (meaning block for
listeners with a strictly lower priority).

This patch simplifies this method to return a bool (with 'false' meaning
don't block, 'true' meaning block for listeners with a lower priority).

This is a simplification, but also a preparation to change the Priority
enum from a plain 'enum' to 'enum class' (because the latter doesn't
allow expressions like 'priority + 1').