AmdFlash: Support continuity check command.
This function allows the program to check for correct connectivity of the flash
chip, and would fail if there are shorts on adjacent address and data pins.

It seems to be specific to the Infineon / Cypress / Spansion S29GL064S, at least
I haven’t seen it being offered elsewhere.