Use MemBuffer instead of std::array on the stack as buffer.
As Wouter analyzed: Commit 39f7692f9618 changed this:
-               std::array<uint32_t, 512 * 256> pixels;
+               std::array<uint32_t, 512 * 256 * 2> pixels; // max size:  512*256*2  or  256*256*4
In other words: it increases the size of a stack-array from 512kB to 1MB.
Apparently on windows the stack is very limited(?), and this change pushes the total stack size over the limit?

So, herewith, we put the buffer on the heap using the MemBuffer class. Let's hope it helps...