WD2793: fix off-by-one error in write-sector command
Internally the WD2793 has 2 data registers. One for the active data byte
that's currently being written to disk. And another to receive the next
to-be-written byte from the CPU. When the active byte is fully written
to disk, the buffered byte is copied to the active byte, and the DRQ
flag is activated again, indicating that the CPU can send the next byte.

But when the last byte is about to be written. That is: the CPU has
already send the last byte, and the previous to-last byte was just
written to disk. The WD2793 copies that last byte to the active byte,
but at this point the DRQ flag should NOT be reactivated.

Before this patch there was a bug, and we did reactivate DRQ, so we
wrongly requested one byte too many from the CPU. This happened to work
anyway for the normal MSX BDOS routines, but it is wrong.

Thanks a lot to Marcos Daniel Blanco Oliveira for reporting this
problem.

After this fix the WD2793 behaves different on write sector commands. So
this might cause a de-sync in old replay files. Though luckily it's not
common for such replays to write to disk.