AmdFlash: Sparse persistence of modified sectors only.
Stores a metadata file in addition to the data. The metadata indicates modified
sectors, and the data contains only those.

Backwards compatible with previous save format; if metadata file is not found
but data file is, assumes all non-write-protected sectors are modified.

Replaces SRAM by own implementation local to AmdFlash.