AmdFlash: Modulo addresses to the size.
Mirroring will be the hardware implementation for unconnected address lines, so
this is better behaviour than returning 0xFF. Preferable to adding an assert,
and `& (flash.size() - 1)` everywhere we call the public interface of AmdFlash.

This is also a means of defensive programming. AmdFlash has nontrivial code,
which has been written with in-range addresses in mind. So by ensuring that it
always is we prevent the unexpected case to begin with.

Change `size()` to return `power_of_two` to improve optimizability.