AmdFlash: Write protect pin improvements.
The WP# pin is not available on all flash memories. Also, depending on the
device, it protects one or two sectors from the bottom or the top.

Additionally, WP# also protects those sectors against chip erase.

Lastly, the status of the write protect pin should not affect the read address
and initial content of the flash memory loaded from persistent storage.
(Even though, since this is loaded on construction, there’s no way to set it.)