DirAsDsk: fix regression since 888c795f959c0d8
Revision 888c795f959c0d8 changed output-parameters into return values.
Part of this implementation also changed reference variables into
by-value variables in helper structs like 'DirScanner' and
'DirEntryForCluster'. This helper struct itsef was passed __by-value__
to scanMsxDirs(). But unfortunately that means that changes to this
struct are done on a copy, and then the caller of scanMsxDirs() doesn't
see these modifications.

I overlooked this bug many times during bug-hunting. But once I found it
the fix was easy: pass the helper struct by reference.

Thanks to grauw and Quibus for helping to review/test this. Together
they gave me a more reliable way to trigger the problem. And with that I
could see in the debugger what was going on.