Fix libstdc++ debug build after changes in AmdFlash
The changes in AmdFlash were the trigger but not the root cause. (I
think) the root cause is a bug in the debug-version of libstdc++??

I added a workaround (replace std::copy() with our implementation). See
comments in the code for more details. Hopefully soon (with half a
year(?), when we can use a newer clang) we can use std::ranges::copy(),
and then we can hopefully remove our workaround.