Add: 'Tools > Simple notes widget'
This implements:
    Request: identify OpenMSX instance on debugger windows
    https://github.com/openMSX/openMSX/issues/1625

This is intentionally kept minimal. A window with _only_ a multi-line
text input field. Nothing is done with this text (except for displaying
it). There can be multiple of these windows open at the same time (each
with their own text). These windows are saved in the openMSX session.
There is intentionally no option to load/save this text from/to a file.
But an easy workaround is to use to regular copy/paste functionality.

The typical use case for this window is to dock into some other window.
And remember that it's possible to hide the title of docked windows.