AmdFlash: Mask fast program address to page size.
According to the datasheets, fast program commands should use addresses within
the same page, that is, for a double byte program the addresses should only
differ in A0, and for a quadruple, A0-A1. It is not specified what happens if
the writes are not within the same page.

In practice on the M29W640GB, the fast program commands all seem to operate with
a page size of 8, meaning all addresses within this page are ok. When writing to
addresses across page boundaries, the definitive page is determined by the last
byte of the command, all other addresses are masked to that same page.