AmdFlash: Change cmd to a std::vector.
So that we no longer have to worry about the maximum command size that needs to
fit in the static vector. The single allocation we save by using static_vector
isn’t worth the hassle IMO.

When buffer program page size is 1 (command not available) it reserves length 6,
which coincides with the next longest command length (chip / sector erase).