RomAscii16X: Remove device factory.
This caused an assertion when opening the “Save state...” menu, due to
inheriting from MSXRom which expects a mappertype tag that was not present.
It’s not really possible to be an MSXRom and a general device at the same time.

Instead the XML now specifies a ROM with mapper type, similar to other devices.

p.s. Clearly the closing tag was not being checked 😅.