Tweak icons for filter (search) and reset (undo).
Instead of the Search concept, with a magnifying glass icon, use a Filter
concept (with a Filter icon). This better reflects the functionality.

Also, instead of the weird reset icon (that is actually a 'share_alt" icon in
the original font), use an undo icon instead. That feels much better to me.

Fixes #1650.