Disable persistent breakpoints (for now?)
See the discussion here:
    https://github.com/openMSX/openMSX/issues/1665

Persistent breakpoints/watchpoint/debug-conditions were added because
the old debugger also had them, but without a clear use-case.

And now they are causing problems related to Tcl scripts that create
break/watchpoints.

Disabled it for now. At least until we get a clear motivation and
solution for how to combine it with scripts.