ImGuiWaveViewer: replace ImGui::Plot{Lines,Histogram}() with our own version
The reason is that the ImGui version has some behavior that we don't
want. For example the popups, and they can't be disabled. Our version can
also be more specific, thus it may be a bit faster (less conditionals in
the inner loop).

The original ImGui::PlotLines() and ImGui::PlotHistogram() functions
weren't that complex (and much less so after our simplications). In fact
the (internal) documentation on those functions has this section:
    // Plot/Graph widgets are not very good.
    // Consider writing your own, or using a third-party one, see:
    // - ImPlot https://github.com/epezent/implot
    // - others https://github.com/ocornut/imgui/wiki/Useful-Extensions
And thus indeed, for now we've written our own version. Later we might
consider using "implot".

--

A future extension could be to re-add popups in the spectrum plot. But
then with meaningful information, e.g. the frequency and power-level.
(For the waveform plot I don't think a popup will be useful).