VDP: expose getMSXPos() in public API
Before this functionality was only exposed via the info-topics:
* msx_y_pos
* msx_x256_pos
* msx_x512_pos
But later we may want to use this info in the ImGui debugger, and then
it's easier when it's directly available as a c++ method. The info
topics are now implemented in terms of this new method.

There were 3 separate info-topics, each with their own (slightly
different) calculation, but there's only 1 new getMSXPos() method:
* getMSXPos() returns both x and y coordinates (in a ivec2).
* Often you're interested in both the x and y coordinates, so a single
  method is more convenient.
* Calculating both at once is more efficient, because part of the
  calculation is shared.
* In case you actually only want x or only y, then the compiler's
  optimizer will most likely be able to optimize the unused part away
  (because getMSXPos() is an inline method).
* The x-coordinate is only exposed in 'narrow' (aka 'screen 7') units.
  But that's fine, the conversion to other modes is trivial.

Thanks to MBilderbeek for starting this patch (I only fine-tuned it a
bit).