DebuggableEditor: add 'forceScroll' in scrollAddr()
The scrollAddr() method would not actually change the scrollbar position
when 'currentAddr' is already equal to the requested address. E.g. (when
emulation is paused) this allows to still manually scroll to a different
position even when the address is given as an expression (e.g. linked to
the HL register).

More in detail: every frame we re-evaluate the address expression and
pass the result to scrollAddr(). This always updates the cursor and
'currentAddress', so the internal state is constantly kept up-to-date.
But setting the scroll-bar position is skipped when 'currentAddress'
did not actually change.

In other words: the internal state is always kept up-to-date, but the
scrollbar position is set only once. Normally that's fine. However there
is a problem when setting the scrollbar position is done (just once)
when the window is not visible (e.g. because it's in a non-active
tab-bar). Apparently such updates get lost. As a solution we force to
update the scrollbar each time the window 'appears' (e.g. when switched
from a non-active to an active tab).

Thanks to Bengalack for reporting this problem:
    https://www.msx.org/forum/msx-talk/openmsx/please-help-testing-upcoming-openmsx-release?page=110
    Post of 14-07-2024, 16:24