MSXMixer::updateStream(), don't needlessly initialize 'mixBuffer'
See here for a use-case that became unexpectedly slower.
    https://github.com/openMSX/openMSX/issues/1708

Background info:
Each time a register in a sound-device is written, we "sync" the
sound-generation. That is: we generate sound right up-to the moment
before the register write, and only then do the write. Indirectly such a
"sync" calls the MSXMixer::updateStream() function.

Inside MSXMixer::updateStream() we have a local (stack) buffer that
holds the to-be-generated audio signal. This buffer contains 8192 x 2
floats. 8192 is the maximum number of samples that will be produced per
call (though in case of very rapid "syncs", this number will often be
zero). And a factor 2 for stereo. This buffer will be overwritten, so
it's not needed to initialize it.

In older openMSX version this buffer was defined as:
    std::array<float, 8192 * 2> mixBuffer;
Later it was changed to:
    std::array<StereoFloat, 8192> mixBuffer;
So instead of manually keeping track of even indices are for the left
channel, odd indices are the right channel. We now have help from the
compiler to keep this straight.

Though 'StereoFloat' was defined like:
    struct SteroFloat {
        float left = 0.0f;
        float right = 0.0f;
    };
Note the "= 0.0f" part. Usually it's a good idea to always initialize
all members. Though in this case this also caused the "mixBuffer" array
above to be (needlessly) initialized. When the MSX program is doing
'OUTI' instructions in a tight loop, this is done ~200k times per
second. So this would write ~12GB/s. And proportionally more during
'fast-forward'. That can indeed cause some slowdowns ;-)

In the "normal" case (when not doing so many sound-device register
writes), the overhead is _much_ less. But still good to avoid it.

Easy to fix. The difficult part was figuring out what was going on.