halfband: switch to cheaper filter
During development of the audio channel viewer (the spectrum part of
it), I switched from a 15 to a 31-point halfband filter, because there
was too much noise in the spectrum. Though that noise was caused by
something else (a bug, I don't remember what exactly), and I forgot to
revert to the cheaper halfband filter.

This saves a few percent host CPU time when many channels are visible
(e.g. all moonosund channels).