SoundDevice::getLastBuffer(): keep track of silence
Some (no all) sound devices have an optimization to indicate channels
that are silent. In this patch we use that information in
getLastBuffer(), we track if the full returned buffer would be silent or
not.

And then in ImGuiWaveViewer we can skip some calculations (e.g. FFT)
when we know the full buffer is silent.