Follow up on previous patch
Some sound device do not report silent channels, an important example is
YM2413NukeYKT. So the previous patch can't help with those.

However in ImGuiWaveViewer we already calculate the power of the signal
(for the VU-meter). If that's zero that also indicates the channel is
silent. And then we can skip the more expensive FFT calculation.