halfband: allow to calculate in-place
The PSG produces sound at ~224kHz (3.57MHz / 16). So 1/7 second takes
31960 samples. We want to reduce that to below 2048 samples (and then
zero-pad to exactly 2048 samples) so that we can pass it to the FFT
function (which requires a power-of-two sized input).

Our current halfband function can roughly cut the buffer size in half
(more in detail: round down to an odd size, then subtract 13, then
divide by 2). So 31960 gets reduced in 4 steps (via 15973, 7980, 3982)
to 1985 samples.

Before this patch we needed space for all these intermediate buffers
(=15973+7980+3982+1985 = 29921 samples). Now we can calculate the
halfband filter in-place, and thus we only need the largest of these
intermediate buffer (=15973 samples), thus almost halving the required
memory space. This improves cache-locality.