ImGuiWaveViewer: replace std::log10() with 'fast_log2()'
For every (active and visible) channel of every sound device in the
'Audio channel viewer', for every frame. And for every bin in the
spectrum (current 1024 bins) we calculate a std::log10(). This was
enough that (on my system when playing a moonsound-wave song) these
calculations took ~3% of total host CPU time.

This patch replaces 'std::log10()' with 'fast_log2()'. The latter can be
calculated much faster, and it can also be auto-vectorized (=SIMD) (I
checked with gcc). After this transformation the calculation is no
longer visible in the profile (takes less than 0.1% CPU).

The difference is that fast_log2() is less accurate than std::log10().
The theoretical maximum error is about 0.25%. That means if we make the
VU-meter widget 1000 pixels wide, it can be off by up to 2 pixels. I
think that's an acceptable trade-off. (And that's the maximum error, I
ran some empirical tests and it was never off by even a single pixel,
but errors of 0.1 pixel (0.01%) were very common).