ImGuiWaveViewer, use cheaper FFT depending on graph size
Before this patch we would always calculate a 2048 point FFT. That gives
1024 frequency bins. Though by default we only draw those 1024 bins in
an array that's only about 200 pixels wide. Meaning 1/5 of the
calculation isn't visible.

In this patch we will dynamically choose between a 256, 512, 1024 or
2048 point FFT. Depening on the size (in pixels) of the graph. Using the
default size this now uses the 512 point version. And with that I
measured a 5x reduction in computation time spend in the FFTReal
function (and that's in line with what we can expect from a O(n *
log(n)) algorithm.

Note that this does not visually look identical. A wider FFT can better
classify the different frequencies in narrower frequency-bins. If there
are fewer bins, instead of having narrow peaks in the spectrum there may
be (slightly) wider blobs. So instead of single-pixel peaks it may now
look like 2 or 3 pixel blobs.