ImGuiWaveViewer: switch from blackman to hamming window
The previous patch talked about narrow peaks vs blobs in the spectrum.
The choice of window-function also has an influence on the width of the
peaks. Initially I picked the 'blackman' window because it has better
suppression of the 'side-lobes'. Though that was at a time when I still
had some other bug that caused more noise, and then this window helped
to suppress that (more).

Now I switched to the 'hamming' window. Compared to the blackman window
this has less suppression of the side lobes (less noise suppression) but
also a more narrow 'main-lobe', and that helps with keeping the blobs in
the spectrum more narrow. Today this is a better trade-off.