Fix crash when using multiple 'debugdevice' extensions at once
I'm not sure that's a useful scenario, but it shouldn't crash. Also by
default both devices will use the same IO ports, but you could copy and
edit the xml file and choose different ports for the 2nd instance.

See here for a discussion:
    https://github.com/openMSX/openMSX/issues/1617
Thanks to 'pvmm' for noticing and reporting the problem.

I've fixed it by using a device-dependant setting name:
* Before the setting (to control where the debug output gets written)
  was called "debugoutput".
* When a 2nd 'debugdevice' extension was used, it tried to reuse that
  same name, but that's not allowed. That triggered an assertion.
* This patch changes the name of the setting to
      <device-name> + " output"
  So for the first 'debugdebug' that will be "Debug Device output", for
  the 2nd it will be "Debug Device (1) output" and so on.
* This does break backwards compatibility. Though I believe this
  settings is very rarely used. And if it's used, it's probably by
  technical people who can deal with this change. So IMHO the extra
  complexity to deal with backwards complexity wasn't worth it.