ReverseManager::getSnapshotTimes(): return a view instead of a vector
Instead of materializing the result of getSnapshotTimes() into a
std::vector<double> we return a view. Then later, when the caller
iterates over that view we (lazily) calculate the result. This way we
never need to allocate memory to store the full result at once.

This temporary memory allocation also showed up high in a 'heaptrack'
profile report.