ImGuiDisassembly: limit vector<> allocations (for labels in address column)
For every label in the address columns, we would allocate (and possibly
re-allocate) a std::vector<std::string_view> to store the list of (best)
label candidates. If there are multiple labels visible, we do this
multiple times per frame. This patch limits that to once per frame (we
reuse the same vector capacity for the different lines).