ImGuiSymbols: avoid creating vectors / string-copies during normal paint()
ImGuiSymbols has a loop over the symbol files. In that loop it draws the
symbols for that file, together with a 'reload' and 'remove' button.
Pressing one of these buttons possibly changes the vector of files. We may
not loop over a changing vector. Therefor we looped over a copy of this
vector. Though that required creating a vector (and copying the strings
inside the elements in that vector).

In this patch we postpone the 'reload' or 'remove' action till after the
drawing loop. So pushing the buttons only remembers what to do.

This saves many temporary heap allocations per frame. Discovered by
'heaptrack', because it was high in the profile.