ImGuiDebugger: scroll disassembly view on 'step_back' in console
See issue:
    When doing a step in the console (e.g. step_back), the Disassembly view doesn't scroll to the current PC
    https://github.com/openMSX/openMSX/issues/1647

The step-forward commands (step, step-over, step-out) all un-break, then
execute some instructions and then break again. This sends a "BREAK"
event, and the disassembly view reacts on this by scrolling to the new PC
position.

The step-back command is different. It creates a new machine instance
(from a previous machine in reverse-history) and execute till the
requested moment in time (the instruction right before the current one).
This is a very similar mechanism as is used for "reverse goto", the
mechanism itself is independent of debugging, so at the end it does not
send a "BREAK" event (technically the event is not send because we never
un-break and re-break). And because of that we did not scroll the
disassembly view.

We could fix this by adding the missing BREAK event. But that that's not
what this patch does. Instead we now also listen to the "MACHINE_LOADED"
event, this gets send whenever we switch to a new machine (so maybe
"MACHINE_SWITCHED" would be a better name?). This fixed the problem that
disassembly view does not scroll on step-back. But in addition it also
makes the view scroll on:
* Switch to new machine instance (the "New machine instance" button in
  the "Select MSX machine" window).
* Click in the reverse-bar.