Fix #1706 Unable to move the main window after unplugging a monitor
On startup we try to restore the position of the main openMSX window.
Though in the scenarion described in this bug report (restore on a now
disconnected monitor) that indeed goes wrong.

Fixed by looping over the available monitors. If the window is not
(partially) visible on any of them, then we restore the window centered
over the primary monitor.