Fix #1633 LED icons and reverse bar move when changing scale factor, but not saving settings
We have a helper called 'AdjustWindowInMainViewPort' used by the OSD
icons and the reversebar, to keep the (borderless) window in a sensible
place when the main-window-size changes (via the 'scale_factor'
setting).

This helper reacts to changes in the (main) viewport size, and then
adjusts the relative position of the window. To do this it compares the
current with the old viewport size.

This works fine when openMSX is running. Though this issue is about
changes in scale_factor when openMSX is _not_ running. Because settings
aren't saved, we restart openMSX with a different scale_factor than when
openMSX was stopped. Note that "settings aren't saved" means the Tcl
settings, not the stuff that's stored in "imgui.ini" (e.g. the position
of the ImGui windows), the latter are always saved.

In this situation the mechanism to compare the old with the new viewport
size breaks down. And then we can't adjust the position of the OSD icons
and reverse bar. The fix is to save the required information in
imgui.ini. So in a way make the situation of switching during normal
run-time and between shutdown/restart the same.