Improve how to obtain the 'Keyboard' object
We used to search for the device called "ppi", dynamic_cast that device
to MSXPPI, and then call MSXPPI::getKeyboard(). This has two problems:
* It doesn't work on non-MSX machines (SVI, CVJOY, SEGA).
* It's expensive (lookup by name) and since that last commit we need
  this every frame.

Now the Keyboard class (un)registers itself in the MSXMotherBoard class.
Normally there is exactly one Keyboard object per machine. But we still
allow for none or more than one to not crash on weird/artificial machine
configs. But querying the Keyboard object just return nullptr (for zero)
or the first object (for 1 or more).