Don't allow to focus the reversebar / OSD-icons
The reversebar and the OSD-icons are implemented as Dear ImGui windows.
By default the title bar of these windows is hidden. But it was still
possible to focus those windows. And the effect is that keyboard input
goes to these windows (and is then ignored) instead of going to the main
openMSX window (for use by the emulated MSX keyboard).

E.g. before this patch, when clicking in the reversebar, or on some icon
(even if the icons is fully faded out) focus would be given to one of those
windows. And then you needed to first click somewhere else in the main
window before MSX keyboard input would work again.

This patch also disables navigation-input "in" and "to" these windows
(when the title is hidden). E.g. Ctrl-Tab now skips these windows.

Background info (this patch doesn't change this): Ctrl-Tab still allows
to navigate to the main menu-bar. This is useful to control this menu
via the keyboard (instead of mouse). To return focus to the main window,
press ESC.