Only render 'real' VDP pixels in the 'extended' border
The real VDP only renders ~568 pixels (512 pixels display-area + left
and right borders). But we map these to 640 host pixels (that's a better
fit with host fullscreen mode etc, and I'm ignoring stuff like
horizontal_stretch etc).

These "extra" pixels are usually just an extension of the left and right
border pixels that would also be rendered by the real VDP. There's one
exception: if you rapidly change the border-color register, you can get
different colors in the border. Before this patch it it was possible to
have a change in color in the extended range of the border. This patch
prevents that. Now the extended pixels are always a repetition of pixels
that the real VDP would also render.

Though the motivation for this patch was completely different:
* During rendering of a single line we assume the screen-mode-width
  doesn't change. E.g. we don't change from screen 5 (256 pixels wide)
  to screen 7 (512 pixels wide). E.g. this allows to store a single
  width per line (this is later used to select a "scale" algorithm for
  that line).
* This assumption is justified: the real VDP only activates the new
  screen mode once per line. And thanks to grauw we know exactly when
  during the line. See https://github.com/openMSX/openMSX/issues/1310.
* So 144 cycles since the start of the sync signal, that's somewhere in
  the middle of the left erase period, well before the left border
  starts.
* Though it turned out that cycle 144, when translated to a position in
  the extended host-line, was no longer before the left host-border.
* So when switching e.g. from a 256 to a 512 pixel wide mode, our
  assumption that the full line is renderer with the same pixel-width is
  no longer true. And it could lead to a few uninitialized pixels.
* In practice this was never a problem. E.g. with the default settings
  for horizontal_stretch these uninitialized pixels weren't even
  visible. But during video recording, soon after boot the first frame
  when switching to screen 6 (for the bootlogo), there were a few
  uninitialized pixels rendered to the .avi file. And valgrind
  rightfully detected this as a UMR.

I found this problem while investigating:
    https://github.com/openMSX/openMSX/issues/1715
though I believe the UMR I found (and fixed) was not the cause of that
issue.