Add 'dropped' files to 'recently-used'
See:
    https://github.com/openMSX/openMSX/issues/1642
Note: this commit explicitly only handles insert via drag-and-drop,
      _not_ via a console command

Media (disk, rom, ...) inserted via drag-and-drop is now added to the
list of recently used files in the corresponding "Media" dialog.

The current implementation parses the (Tcl) command to re-construct a
'MediaItem' object. For the code in 'ImGuiMedia' that's a detour: we've
just formatted a 'MediaItem' into a command. But is useful for the (new)
code that handles drag-and-drop.

An alternative (to parsing the command) could have been to, in the
drag-and-drop code, not only construct a command but also a 'MediaItem'.
But then we'd better also move the drag-and-drop code from
'ImGuiManager' to 'ImGuiMedia'.