CPUClock: fix possible crash right after changing the CPU frequency
See:
    https://github.com/openMSX/openMSX/issues/1728

During CPU emulation we _very_ frequently (~a million times per second) need
to check when to exit the CPU-emulation loop. So it's important that
this check is super fast. We managed to reduce that check to
    "if (remaining < 0) { .."
So (on most host CPUs including x86, ARM, ...)
* (at most) a single (32-bit) load
* a test or compare instruction
* a conditional branch.
In particular we do not need to compare two (64-bit) values (and also we
don't need to test a flag). See the comment block above
CPUClock::setLimit() (CPUClock.hh:74) for lots more details.

The trick was to transform a comparison between two (64-bit) timestamps
into a single (32-bit) "are there any remaining clock-cycles"-check.

Though the duration of "a clock-cycle" changes after changing the CPU
frequency. So right after changing the "Z80_freq" setting (which
triggers CPUClock::setFreq()) the bookkeeping in CPUClock could be
wrong. And in some cases that could lead to an assertion failure.

Fixed by calling "disableLimit()" (once when freq changes). This maybe
causes one spurious exit from the CPU emulation loop. But that's
totally fine.