Optimize Display::updateZ() / fix theoretical problem
In the Display class we keep a vector of layer object, sorted from back
to front (from lower to higher Z-coordinate). When the 'videosource'
setting is changed, the z-coordinate of some layers changes. This is
done via the updateZ() function.

Before we implemented updateZ() as:
* Remove the layer from the old position. This moves all elements past
  the old position 1 place to the front.
* Re-insert it in the new position. This moves all elements part this new
  position 1 place to the back.
So possibly some elements (at the tail) needlessly moved 1 place to
the front and then immediately after 1 place back again, so to their
original location.

Now we implement updateZ() as:
* Locate the old and the new position.
* And (only if needed) perform a std::rotate().
This does the minimal number of assignments (no unnecessary ones).

In practice the vector only contains a few elements (typically 6 or
fewer). So the old implementation wasn't too bad. Though the actual
motivation for this patch was the following:

While drawing we iterate over all layers in the vector. One of these
layer is the ImGuiLayer. If you select "menu bar > settings > video >
videosource" then this may indirectly also change the Z-coordinate of
some of the layers. In other words it reorders the elements in the
vector we're iterating over.

In practice this isn't a problem: the netto effect of remove followed by
re-insert (at a different position) is only a shift of some of the
elements. The vector size has remained the same and it also hasn't been
reallocated. So the worst that can happen is that, when we continue
iterating over the vector, we draw other layers. But in practice even
that isn't possible because the ImGui layer has a higher Z-value than
any of the reorderable layers (IOW it's the last layer that gets drawn).

However when we compile with the DEBUG-STL enabled. The instrumentation
of vector::insert() DOES invalidate the iterator in our loop. The
instrumentation isn't smart enough to recognize the special remove +
re-insert pattern. So we do trigger a (false-positive) error.

I don't mind this false positive at all, because it does point to an
inefficiency. And that's fixed now, together with the error.