Moonsound: fix (debug-)read of wave-register 0xf8 (FM-mix-level)
Everything worked as it should, except querying the reset-value of
register 0xf8 (so before it was ever written).

More in detail: the initial (after reset) FM-mix-levels should be -9dB.
This was correctly emulated like this. Only when (debug-) reading this
register value, it incorrectly showed the value corresponing to levels
0dB.

Thanks to 'Bengalack' for reporting!
See https://www.msx.org/forum/msx-talk/openmsx/please-help-testing-upcoming-openmsx-release?page=114
 post of 08-09-2024, 09:56