Throttle ImGui messages
When some Tcl script is generating warnings at a too high rate the GUI
could become unusable. It could even be so bad that the only way to
recover was to kill openMSX.

For example:
   debug set_bp 0x38 bla
("bla" is an invalid Tcl expression, and this gets evaluated every IRQ,
50 times per second)

In this patch we throttle the messages if there are too many of them in
a too short timespan. At the moment I picked:
 * more than 5 within a second
 * or more than 10 within a minute.
All the latest messages are still shown (same as before) but during
throttling we don't re-focus the message window or show any (modal)
popups.

This patch is just a proof-of-concepts. The detail of how exactly we
throttle can still be fine-tuned.