More direct way to test if main menu bar is hovered
See
  https://github.com/openMSX/openMSX/issues/1756
  Menu bar doesn't (reliably) appear on full screen in Windows, in contrast to Linux #1756

We only tested whether the main menu bar was hovered in an very indirect
way: check if any window is hovered. But this relies on an
implementation detail: the main menu bar is internally implemented as an
ImGui window.

In this patch we more directly check the mouse coordinates. Maybe that
solves the windows/fullscreen issue???

But as a bonus this approach allows to _fully_ fade out the menu bar.
Before we only faded out till alpha=0.001. In practice this is
invisible, but because it's not exactly zero, it technically still gets
drawn (and that was needed for the (old) window-is-hovered check).